Text Changes
Text looks good. I'm a little concerned that Jeff will not find the story a compelling one for user training. It may need to be rooted in an actual scenario. Something like: If you are working with a large reports and need to select filters among many possible filters, you may find it useful to simplify by designating specific lists of filter options for certain key parameters, such as customer accounts. Interject's jDropdown formula does this by...

I think the key is building that context.